finish preparing
