+++ src/App.vue
@@ -1,11 +1,13 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
-import AppHeader from '@/components/AppHeader.vue'
+import TheHeader from './components/TheHeader.vue'
+import TheFooter from './components/TheFooter.vue'
 </script>
 
 <template>
-  <AppHeader></AppHeader>
+  <TheHeader></TheHeader>
   <RouterView />
+  <TheFooter></TheFooter>
 </template>
 
 <style scoped></style>
