+++ src/assets/base.css
@@ -1,17 +1,14 @@
 /* color palette from <https://github.com/vuejs/theme> */
 
-
 /* :root {
 
 } */
 
 /* semantic color variables for this project */
 
-
 /* :root {
 
 } */
-
 
 /* @media (prefers-color-scheme: dark) {
   :root {
@@ -28,18 +25,32 @@
   font-weight: normal;
 }
 
+html {
+  height: 100%;
+  margin-right: calc(-1 * (100vw - 100%));
+  overflow-x: hidden;
+}
 body {
-  min-height: 100vh;
+  font-family: Roboto, sans-serif;
+}
+body,
+#app {
+  height: 100%;
+}
+#app {
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+  height: 100%;
 }
 
-
 .container {
-  width: 100vw;
-  margin: 0 auto;
+  width: calc(100% - 10px);
+  margin: 0 5px;
 }
 @media (min-width: 768px) {
   .container {
     width: 750px;
+    margin: 0 auto;
   }
   /* main {
     display: flex;
