+++ src/components/UI/index.ts
@@ -1,3 +1,3 @@
-import CustomButton from '@/components/UI/CustomButton.vue';
-import CustomInput from '@/components/UI/CustomInput.vue';
-export default [ CustomInput, CustomButton] 
+import CustomButton from '@/components/UI/CustomButton.vue'
+import CustomInput from '@/components/UI/CustomInput.vue'
+export default [CustomInput, CustomButton]
