+++ src/components/__tests__/HelloWorld.spec.ts
@@ -1,8 +1,8 @@
 // import { describe, it, expect } from 'vitest'
-// 
+//
 // import { mount } from '@vue/test-utils'
 // import HelloWorld from '../MainComponent.vue'
-// 
+//
 // describe('HelloWorld', () => {
 //   it('renders properly', () => {
 //     const wrapper = mount(HelloWorld, { props: { msg: 'Hello Vitest' } })
