+++ src/directives/index.ts
@@ -1,4 +1,4 @@
-import vFocus from "@/directives/vFocus";
-import vIntersection from "@/directives/vIntersection";
+import vFocus from '@/directives/vFocus'
+import vIntersection from '@/directives/vIntersection'
 
-export default [vFocus,vIntersection]
+export default [vFocus, vIntersection]
