+++ src/directives/vFocus.ts
@@ -1,7 +1,6 @@
-const name = "focus"
+const name = 'focus'
 const mounted = (el: any) => el.focus()
 
-
-const directive  = {name, mounted}
+const directive = { name, mounted }
 
 export default directive
