+++ src/main.ts
@@ -5,17 +5,17 @@ import { createPinia } from 'pinia'
 
 import App from './App.vue'
 import router from './router'
-import ui from "@/components/UI";
-import directives from "@/directives"
+import ui from '@/components/UI'
+import directives from '@/directives'
 
 const app = createApp(App)
 
 app.use(createPinia())
 app.use(router)
-ui.forEach(component=>{
-    app.component(component.name,component)
+ui.forEach((component) => {
+  app.component(component.name, component)
 })
-directives.forEach((directive)=>{
-    app.directive(directive.name,directive)
+directives.forEach((directive) => {
+  app.directive(directive.name, directive)
 })
 app.mount('#app')
