+++ src/stores/counter.ts
@@ -8,6 +8,6 @@ export const useCounterStore = defineStore('counter', {
     },
     decrement() {
       this.count--
-    },
-  },
+    }
+  }
 })
