+++ src/views/HomeView.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-
 import HomeComponent from '@/components/HomeComponent.vue'
 </script>
