+++ src/views/TodoView.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import MainComponent from '@/components/HomeComponent.vue'
+</script>
+
+<template>
+  <main>
+    <MainComponent />
+  </main>
+</template>
